requirements.txt
asgiref==3.4.1
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cssselect2==0.4.1
Django==3.2.6
django-appconf==1.0.5
django-ckeditor==6.1.0
django-image-cropping==1.7
django-js-asset==1.2.2
django-social-share==2.2.1
django-storages==1.12.3
django-taggit==1.2.0
easy-thumbnails==2.8.1
idna==3.3
isodate==0.6.1
lxml==4.8.0
msrest==0.6.21
oauthlib==3.1.1
Pillow==8.3.2
psycopg2-binary==2.9.3
pycparser==2.21
python-decouple==3.4
pytz==2021.1
reportlab==3.6.7
reqs==0.0.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.1
svglib==1.2.1
tinycss2==1.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==5.3.0
